Add logging, use dry run for now

+++ service/src/main/kotlin/fi/espoo/evaka/attachment/AttachmentService.kt
@@ -83,8 +83,15 @@ WHERE id = ${bind(id)}
         }
     }
 
-    fun scheduleOrphanAttachmentDeletion(tx: Database.Transaction, clock: EvakaClock) {
+    fun scheduleOrphanAttachmentDeletion(
+        tx: Database.Transaction,
+        clock: EvakaClock,
+        dryRun: Boolean
+    ) {
         val ids = tx.getOrphanAttachments(olderThan = clock.now().minusDays(1))
-        asyncJobRunner.plan(tx, ids.map { AsyncJob.DeleteAttachment(it) }, runAt = clock.now())
+        logger.info("Scheduling deletion for ${ids.size} orphan attachments")
+        if (!dryRun) {
+            asyncJobRunner.plan(tx, ids.map { AsyncJob.DeleteAttachment(it) }, runAt = clock.now())
+        }
     }
 }
